Time: 313 ms (53.06%), Space: 127.2 MB (25.00%) - LeetHub

+++ 1695-maximum-erasure-value/1695-maximum-erasure-value.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int maximumUniqueSubarray(vector<int>& nums) {
-        long long int ans =0;
+         int ans =0;
         int i = 0, j=0;
-        long long int sum =0;
+         int sum =0;
         unordered_set<int> res;
         while(j<nums.size() && i<=j ){
             if(res.find(nums[j]) == res.end()){
